# src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: TopAgents API
  description: |
    ## TopAgents.ca — Real Estate Agent Review & Referral Platform

    ### Authentication
    Protected endpoints require a `Bearer` JWT token in the `Authorization` header.
    Get tokens via `POST /api/v1/auth/login`. Refresh via `POST /api/v1/auth/refresh`.

    ### Rate Limiting
    - General: 100 req / 15 min
    - Auth endpoints: 10 req / 15 min
    - Search: 60 req / 1 min

    ### Pagination
    All list endpoints accept `?page=1&limit=20`.
    Response includes `meta: { total, page, limit, totalPages, hasNext, hasPrev }`.

    ### Response Envelope
    ```json
    { "success": true, "data": { ... } }
    ```
    ```json
    { "success": false, "error": { "code": 422, "message": "Validation failed", "details": [] } }
    ```
  version: 1.0.0
  contact:
    name: TopAgents Dev Team
    email: dev@topagents.ca
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.topagents.ca/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    Referral:
      type: object
      properties:
        id:
          type: integer
        sender_id:
          type: integer
        receiver_id:
          type: integer
        client_name:
          type: string
        property_type:
          type: string
        transaction_type:
          type: string
        budget_min:
          type: integer
          nullable: true
        budget_max:
          type: integer
          nullable: true
        location:
          type: string
        timeline:
          type: string
          nullable: true
        referral_fee_percent:
          type: number
          nullable: true
        status:
          type: string
          enum: [PENDING, ACCEPTED, DECLINED, IN_PROGRESS, COMPLETED, CANCELLED]
        message:
          type: string
          nullable: true
        response_message:
          type: string
          nullable: true
        sale_price:
          type: integer
          nullable: true
        commission_amount:
          type: integer
          nullable: true
        referral_fee_paid:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        responded_at:
          type: string
          format: date-time
          nullable: true
        accepted_at:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
    ReferralCreate:
      type: object
      required: [receiver_id, client_name, transaction_type, location]
      properties:
        receiver_id:
          type: integer
        client_name:
          type: string
          maxLength: 200
        client_email:
          type: string
          format: email
        client_phone:
          type: string
        property_type:
          type: string
          enum: [residential, commercial, investment, land]
        transaction_type:
          type: string
          enum: [buy, sell, rent]
        budget_min:
          type: integer
        budget_max:
          type: integer
        location:
          type: string
          maxLength: 200
        timeline:
          type: string
        notes:
          type: string
        referral_fee_percent:
          type: number
          minimum: 0
          maximum: 50
        message:
          type: string
    Lead:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: integer
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        looking_for:
          type: string
          enum: [buy, sell, rent, invest]
        property_type:
          type: string
          nullable: true
        budget_min:
          type: integer
          nullable: true
        budget_max:
          type: integer
          nullable: true
        location:
          type: string
          nullable: true
        status:
          type: string
          enum: [NEW, CONTACTED, QUALIFIED, CONVERTED, CLOSED]
        source:
          type: string
        created_at:
          type: string
          format: date-time
        contacted_at:
          type: string
          format: date-time
          nullable: true
    LeadCreate:
      type: object
      required: [agent_id, name, looking_for]
      properties:
        agent_id:
          type: integer
        name:
          type: string
          maxLength: 200
        email:
          type: string
          format: email
        phone:
          type: string
        message:
          type: string
        looking_for:
          type: string
          enum: [buy, sell, rent, invest]
        property_type:
          type: string
        budget_min:
          type: integer
        budget_max:
          type: integer
        location:
          type: string
        timeline:
          type: string
        source:
          type: string
          enum: [topagents_search, agent_profile, city_page, quiz, direct]
    Tier:
      type: object
      properties:
        tier_name:
          type: string
        min_points:
          type: integer
        max_points:
          type: integer
          nullable: true
        tier_order:
          type: integer
        min_reviews:
          type: integer
        min_avg_rating:
          type: number
        min_platforms_connected:
          type: integer
        color_hex:
          type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /agents/{agentId}/points:
    get:
      tags: [Points & Tiers]
      summary: Get point transaction history for an agent
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
        - in: query
          name: year
          schema:
            type: integer
            example: 2026
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated point transactions with breakdown by action type
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/{agentId}/tier:
    get:
      tags: [Points & Tiers]
      summary: Get agent tier progress and all tier thresholds
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Current tier, year_points, next tier, and all tier definitions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      current_tier:
                        type: string
                      year_points:
                        type: integer
                      total_points:
                        type: integer
                      next_tier:
                        $ref: '#/components/schemas/Tier'
                      points_to_next:
                        type: integer
                      all_tiers:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tier'
  /agents/{agentId}/connections:
    get:
      tags: [Points & Tiers]
      summary: List connected review platforms for an agent
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Array of platform connections with sync status

    post:
      tags: [Points & Tiers]
      summary: Connect a review platform (awards 75pts first, 35pts each additional)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform:
                  type: string
                  enum: [google, rew, ratemyagent, ratemyagentcom, facebook, zillow, yelp]
                platform_account_id:
                  type: string
                credentials:
                  type: object
                  description: OAuth tokens — stored encrypted in JSONB
      responses:
        '200':
          description: Platform connected and points awarded
  /agents/{agentId}/connections/{platform}:
    delete:
      tags: [Points & Tiers]
      summary: Disconnect a review platform
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: integer
        - in: path
          name: platform
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Platform disconnected

  # ─── REVIEWS ───────────────────────────────────────────────────────────────
...
  /referrals:
    get:
      tags: [Referrals]
      summary: List referrals for the authenticated agent
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: direction
          schema:
            type: string
            enum: [sent, received, all]
            default: all
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, ACCEPTED, DECLINED, IN_PROGRESS, COMPLETED, CANCELLED]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated referrals
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Referral'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags: [Referrals]
      summary: Send a new referral to another agent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReferralCreate'
      responses:
        '201':
          description: Referral created with PENDING status
        '400':
          description: Cannot refer to yourself
        '404':
          description: One or both agents not found
  /referrals/{referralId}:
    get:
      tags: [Referrals]
      summary: Get referral details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: referralId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Referral with sender and receiver info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Referral'
        '404':
          $ref: '#/components/responses/NotFound'
  /referrals/{referralId}/respond:
    patch:
      tags: [Referrals]
      summary: Accept, decline, or counter-offer a referral
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: referralId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [accept, decline, counter]
                response_message:
                  type: string
                counter_fee_percent:
                  type: number
                  description: Required when action is counter
      responses:
        '200':
          description: Referral status updated
        '400':
          description: Already responded or invalid status transition
  /referrals/{referralId}/complete:
    patch:
      tags: [Referrals]
      summary: Mark referral as completed — awards 80 points to both agents
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: referralId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sale_price, commission_amount, referral_fee_paid]
              properties:
                sale_price:
                  type: integer
                  description: Final sale price in cents
                  example: 85000000
                commission_amount:
                  type: integer
                  description: Total commission in cents
                  example: 2125000
                referral_fee_paid:
                  type: integer
                  description: Referral fee paid in cents
                  example: 531250
      responses:
        '200':
          description: Referral completed — 80 points awarded to both agents

  # ─── LEADS ─────────────────────────────────────────────────────────────────
  /leads:
    post:
      tags: [Leads]
      summary: Submit a lead — consumer contacts an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadCreate'
      responses:
        '201':
          description: Lead created and agent notified
        '404':
          description: Agent not found
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [Leads]
      summary: List leads for the authenticated agent
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [NEW, CONTACTED, QUALIFIED, CONVERTED, CLOSED]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
  /leads/{leadId}/status:
    patch:
      tags: [Leads]
      summary: Update lead pipeline status — awards points for fast first response
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: leadId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [NEW, CONTACTED, QUALIFIED, CONVERTED, CLOSED]
                notes:
                  type: string
      responses:
        '200':
          description: Status updated. 20pts if responded within 1hr, 10pts within 2hr.
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── CITIES ────────────────────────────────────────────────────────────────
...